% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nc.R
\name{add_nc}
\alias{add_nc}
\title{Add nc}
\usage{
add_nc(path)
}
\arguments{
\item{path:}{A character string representing the file path.}
}
\value{
: A character string representing the modified file path with "data.nc" appended to it.
}
\description{
: This function takes a path as input and returns a modified path by appending "data.nc" to it.
}
\examples{
add_nc("my_folder/")  # Returns "my_folder/data.nc"
add_nc("/path/to/file.txt")  # Returns "/path/to/file.txtdata.nc"
add_nc("")  # Returns "data.nc"

}
