% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_t_range.R
\name{add_t_range}
\alias{add_t_range}
\title{add t range}
\usage{
add_t_range(path, min_date, max_date, dim_t = "T")
}
\arguments{
\item{path}{The base path to which the time range will be added.}

\item{min_date}{The minimum date of the time range.}

\item{max_date}{The maximum date of the time range.}

\item{dim_t}{The dimension of time to be included in the range (default: "T").}
}
\value{
A string representing the path with the time range added.
}
\description{
This function generates a string representing a time range to be added to a given path.
}
\examples{
# Example 1: Generate a time range string with default dimension
#path <- "http://example.com/"
#min_date <- lubridate::ymd("2023-01-01")
#max_date <- lubridate::ymd("2023-12-31")
#t_range <- add_t_range(path, min_date, max_date)


}
