% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_xy_area_range.R
\name{add_xy_area_range}
\alias{add_xy_area_range}
\title{add xy area range}
\usage{
add_xy_area_range(
  path,
  min_lon,
  max_lon,
  min_lat,
  max_lat,
  dim_x = "X",
  dim_y = "Y"
)
}
\arguments{
\item{path}{The base path to which the XY area range will be added.}

\item{min_lon}{The minimum longitude of the area range.}

\item{max_lon}{The maximum longitude of the area range.}

\item{min_lat}{The minimum latitude of the area range.}

\item{max_lat}{The maximum latitude of the area range.}

\item{dim_x}{The dimension for longitude (default: "X").}

\item{dim_y}{The dimension for latitude (default: "Y").}
}
\value{
A string representing the path with the XY area range added.
}
\description{
A function that generates a string representing an XY area range to be added to a given path.
}
\examples{

# Example 2: Generate an XY area range string with custom dimensions
#path <- "http://example.com"
#min_lon <- -90
#max_lon <- -80
#min_lat <- 30
#max_lat <- 40
xy_range <- add_xy_area_range(path, min_lon, max_lon, min_lat, max_lat, dim_x = "LON", dim_y = "LAT")
}
