% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_xy_point_range.R
\name{add_xy_point_range}
\alias{add_xy_point_range}
\title{Create a File Path for XY Point Range Data}
\usage{
add_xy_point_range(path, min_lon, min_lat, dim_x = "X", dim_y = "Y")
}
\arguments{
\item{path}{(character) The base path where the data file will be stored.}

\item{min_lon}{(numeric) The minimum longitude value for the range.}

\item{min_lat}{(numeric) The minimum latitude value for the range.}

\item{dim_x}{(character) The name of the X dimension. (Default: "X")}

\item{dim_y}{(character) The name of the Y dimension. (Default: "Y")}
}
\value{
The generated file path for the XY point range data as a character string.
}
\description{
This function generates a file path for XY point range data based on the provided parameters.
}
\examples{
add_xy_point_range("data", -90, 30, "X", "Y")

}
