% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancor_coef.R
\name{cancor_coef}
\alias{cancor_coef}
\title{Cancor coef}
\usage{
cancor_coef(object)
}
\arguments{
\item{object:}{An object containing "xcoef" and "ycoef" elements.}
}
\value{
: A subset of the object containing only the "xcoef" and "ycoef" elements.
}
\description{
: This function takes an object as input and returns the "xcoef" and "ycoef" elements from the object.
}
\examples{

data <- list(xcoef = c(0.5, 0.3, -0.2), ycoef = c(0.8, -0.4, 0.6))
cancor_coef(data)
# Returns:
# $xcoef
# [1] 0.5 0.3 -0.2
#
# $ycoef
# [1] 0.8 -0.4 0.6
}
