% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic_details.R
\name{climatic_details}
\alias{climatic_details}
\title{Climatic Details}
\usage{
climatic_details(
  data,
  date,
  elements = ...,
  stations,
  order = FALSE,
  day = TRUE,
  month = FALSE,
  year = FALSE,
  level = FALSE
)
}
\arguments{
\item{data}{A dataset containing climatic data.}

\item{date}{The date variable in the dataset.}

\item{elements}{A character vector specifying the climatic elements to analyze.}

\item{stations}{A character vector specifying the stations to analyze.}

\item{order}{A logical value indicating whether the resulting tables should be ordered.}

\item{day}{A logical value indicating whether to include information at the day level.}

\item{month}{A logical value indicating whether to include information at the month level.}

\item{year}{A logical value indicating whether to include information at the year level.}

\item{level}{A logical value indicating whether to include the level information in the output.}
}
\value{
A data frame containing climatic details such as start and end dates of missing values and the count of missing values for each element at the specified levels.
}
\description{
This function extracts climatic details from a given dataset based on specified parameters such as date, elements, stations, and level. It provides information about the duration and frequency of missing values for each element at different levels (day, month, year).
}
\examples{
#climatic_details(data = climatic_data, 
#                 date = "Date", 
#                 elements = c("Temperature", "Precipitation"), 
#                 stations = c("Station1", "Station2"), 
#                 order = TRUE, 
#                 day = TRUE, 
#                 month = FALSE, 
#                 year = TRUE, 
#                 level = TRUE)
}
