% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic_missing.R
\name{climatic_missing}
\alias{climatic_missing}
\title{climatic_missing}
\usage{
climatic_missing(
  data,
  date,
  elements = ...,
  stations,
  start = TRUE,
  end = FALSE
)
}
\arguments{
\item{data}{A data frame containing climatic data.}

\item{date}{A variable representing the date or time in the data frame.}

\item{elements}{A character vector specifying the climatic elements to be analyzed. Defaults to all elements in the data frame.}

\item{stations}{A character vector specifying the stations to be analyzed. If not provided, the analysis will be performed for all stations in the data frame.}

\item{start}{A logical value indicating whether to determine the start date of missing data for each station and element. If set to TRUE, the function finds the first date with missing values. If set to FALSE, the function uses the first date in the specified time period.}

\item{end}{A logical value indicating whether to determine the end date of missing data for each station and element. If set to TRUE, the function finds the last date with missing values. If set to FALSE, the function uses the last date in the specified time period.}
}
\value{
A data frame containing the summary of missing data, including the start and end dates, the total number of missing values, the percentage of missing values, and the number of complete years for each station and element.
}
\description{
This function calculates the summary of missing data in climatic variables for different stations and elements within a specified time period. It provides information on the start and end dates of missing data, the total number of missing values, and the percentage of missing values for each station and element. Additionally, it counts the number of complete years (without any missing values) for each station and element.
}
\examples{
# Example 1: Calculate missing data summary for all elements and stations
summary_data <- climatic_missing(data = climatic_data, date = "Date")

# Example 2: Calculate missing data summary for specific elements and stations
summary_data <- climatic_missing(data = climatic_data, date = "Date", elements = c("Temperature", "Precipitation"), stations = c("Station1", "Station2"))

# Example 3: Calculate missing data summary with custom start and end dates
summary_data <- climatic_missing(data = climatic_data, date = "Date", start = FALSE, end = TRUE)

# Example 4: Calculate missing data summary without including station information
summary_data <- climatic_missing(data = climatic_data, date = "Date", stations = NULL)


}
