% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_columns.R
\name{compare_columns}
\alias{compare_columns}
\title{compare columns
Compare two columns and provide information about their values}
\usage{
compare_columns(
  x,
  y,
  use_unique = TRUE,
  sort_values = TRUE,
  firstnotsecond = TRUE,
  secondnotfirst = TRUE,
  display_intersection = FALSE,
  display_union = FALSE,
  display_values = TRUE
)
}
\arguments{
\item{x}{The first column to compare.}

\item{y}{The second column to compare.}

\item{use_unique}{If TRUE, remove duplicate values from both columns before comparison. Default is TRUE.}

\item{sort_values}{If TRUE, sort the values in both columns before comparison. Default is TRUE.}

\item{firstnotsecond}{If TRUE, display the values that appear in the first column but not in the second column. Default is TRUE.}

\item{secondnotfirst}{If TRUE, display the values that appear in the second column but not in the first column. Default is TRUE.}

\item{display_intersection}{If TRUE, display the values that appear in both columns. Default is FALSE.}

\item{display_union}{If TRUE, display the values that appear in either column. Default is FALSE.}

\item{display_values}{If TRUE and the columns contain the same values, display the values. Default is TRUE.}
}
\value{
None.
}
\description{
compare columns
Compare two columns and provide information about their values
}
\examples{
x <- 1:10
y <- 5:14
compare_columns(x, y)

}
