% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_sum.R
\name{consecutive_sum}
\alias{consecutive_sum}
\title{Consecutive Sum}
\usage{
consecutive_sum(x, initial_value = NA)
}
\arguments{
\item{x}{A numeric vector.}

\item{initial_value}{A numeric value representing the initial sum value. Default is NA.}
}
\value{
A numeric vector with the consecutive sum values.
}
\description{
This function calculates the consecutive sum of a given vector.
}
\examples{
consecutive_sum(c(1, 2, 3, 4, 5))
# Output: 1 3 6 10 15

consecutive_sum(c(1, NA, 3, 0, 5), initial_value = 10)
# Output: 11 NA 14 14 19
}
