% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_SST.R
\name{convert_SST}
\alias{convert_SST}
\title{Convert SST Data}
\usage{
convert_SST(datafile, data_from = 5)
}
\arguments{
\item{datafile}{A data file containing SST data.}

\item{data_from}{The row index from where the SST data starts in the data file. Default is 5.}
}
\value{
A list containing the converted data frame and the latitude-longitude data frame.
}
\description{
This function converts SST  data from a given file into a structured format. It extracts information such as start year, end year, duration, longitude, latitude, and SST values from the data file. The function then creates a data frame with the extracted information and returns it along with the latitude-longitude data frame.
}
\examples{
# Example usage:
# Assuming the datafile is "sst_data.csv" and the SST data starts from row 5
#converted_data <- convert_SST("sst_data.csv", data_from = 5)

# Access the converted data frame and latitude-longitude data frame
#my_data <- converted_data[[1]]
#lat_lon_df <- converted_data[[2]]

# Print the converted data frame
}
