% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dec_deg.R
\name{convert_to_dec_deg}
\alias{convert_to_dec_deg}
\title{convert to dec deg}
\usage{
convert_to_dec_deg(dd, mm = 0, ss = 0, dir)
}
\arguments{
\item{dd}{The degrees component of the coordinate.}

\item{mm}{The minutes component of the coordinate (default: 0).}

\item{ss}{The seconds component of the coordinate (default: 0).}

\item{dir}{The direction of the coordinate, indicating whether it is east (E), west (W), north (N), or south (S).}
}
\value{
The converted coordinate in decimal degrees.
}
\description{
This function converts coordinates given in degrees, minutes, and seconds format to decimal degrees.
}
\examples{
convert_to_dec_deg(45, 30, 30, "N") # Returns 45.508333
convert_to_dec_deg(122, 10, 0, "W") # Returns -122.166667
}
