% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_yy_to_yyyy.R
\name{convert_yy_to_yyyy}
\alias{convert_yy_to_yyyy}
\title{Convert Two-Digit Year to Four-Digit Year}
\usage{
convert_yy_to_yyyy(x, base)
}
\arguments{
\item{x}{A numeric vector of two-digit years.}

\item{base}{The base year used as a reference for the conversion.}
}
\value{
A numeric vector of four-digit years.
}
\description{
This function converts a two-digit year to a four-digit year based on a given base year. It adds 2000 to the two-digit year if it is less than or equal to the base year, otherwise, it adds 1900 to the two-digit year.
}
\examples{
# Example usage:
# Convert two-digit years to four-digit years based on the base year 1990
years <- c(92, 98, 04, 88)
converted_years <- convert_yy_to_yyyy(years, base = 1990)
print(converted_years)
# Output: 1992 1998 2004 1988
}
