% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_av_packs.R
\name{create_av_packs}
\alias{create_av_packs}
\title{Create Available Packages Data Frame}
\usage{
create_av_packs()
}
\value{
A data frame containing the list of available packages.
}
\description{
This function retrieves the list of available packages from the specified repository and stores it in a data frame.
}
\examples{
create_av_packs()
}
