% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_to_dms.R
\name{dd_to_dms}
\alias{dd_to_dms}
\title{Convert decimal degrees to DMS format}
\usage{
dd_to_dms(x, lat)
}
\arguments{
\item{x}{Numeric value representing the decimal degree to be converted.}

\item{lat}{Logical value indicating whether the conversion is for latitude (TRUE) or longitude (FALSE).}
}
\value{
A character string representing the input decimal degree in DMS format.
The format is "DD MM SS Dir", where DD represents degrees, MM represents minutes,
SS represents seconds, and Dir represents the direction (N, S, E, or W).
The degrees, minutes, and seconds are zero-padded to two digits each.
}
\description{
Converts decimal degrees to degrees, minutes, and seconds (DMS) format.
}
\examples{
dd_to_dms(37.7749, 'TRUE')
# Output: "37 46 29 N"

dd_to_dms(-122.4194, 'FALSE')
# Output: "122 25 10 W"
}
