% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dekade.R
\name{dekad}
\alias{dekad}
\title{Get the dekad component of a date-time object}
\usage{
dekad(date)
}
\arguments{
\item{date}{A date-time object}
}
\value{
a numerical vector of dekad objects corresponding to date variable.
}
\description{
Convert a date or date-time object to a yearly dekad (10-day period)
}
\examples{
dekad(as.Date("2020/12/25"))
dekad(as.Date("1999/01/01"))
}
