% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_unused_levels.R
\name{drop_unused_levels}
\alias{drop_unused_levels}
\title{Drop Unused Levels}
\usage{
drop_unused_levels(data, columns)
}
\arguments{
\item{data}{A data frame.}

\item{columns}{A character vector specifying the column names in \code{dat} from which unused levels should be dropped.}
}
\value{
A modified version of the input data frame \code{data} with unused levels dropped from the specified columns.
}
\description{
This function drops unused levels from specified columns in a data frame.
}
\examples{
# Create a data frame with a factor column
df <- data.frame(
  x = factor(c("A", "B", "C", "D")),
  y = c(1, 2, 3, 4)
)

# Drop unused levels from the 'x' column
df_modified <- drop_unused_levels(df, "x")
levels(df_modified$x)
# Output: "A" "B" "C" "D"

# Drop unused levels from multiple columns
df_modified <- drop_unused_levels(df, c("x", "y"))
levels(df_modified$x)
# Output: "A" "B" "C" "D"

levels(df_modified$y)
# Output: 1 2 3 4 


}
