% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_cases.R
\name{duplicated_cases}
\alias{duplicated_cases}
\title{Duplicated Cases}
\usage{
duplicated_cases(col_name, ignore = NULL, tolerance = 0.01)
}
\arguments{
\item{col_name}{A vector representing the column for which duplicated cases need to be identified.}

\item{ignore}{A vector of values to be ignored when identifying duplicated cases. These values will not be considered as duplicates.}

\item{tolerance}{A numeric value representing the tolerance level for numeric columns. If the absolute difference between two consecutive numeric values is within this tolerance, they are considered duplicates.}
}
\value{
A vector of the same length as the input column, where each element represents a unique identifier for each case. Duplicated cases are assigned the same identifier.
}
\description{
This function identifies duplicated cases in a column and assigns a unique identifier to each duplicate case. The function supports both numeric and non-numeric columns.
}
\examples{
#' # Example 1: Numeric column
col <- c(1, 2, 2.01, 3, 4, 4.005, 4.01, 4.02, 4.03)
duplicated_cases(col)
# Output: 1 1 2 1 1 2 3 4 5

# Example 2: Non-numeric column
col <- c("A", "A", "B", "C", "C", "D", "E")
duplicated_cases(col)
# Output: 1 1 1 1 2 1 1


}
