% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_count_index.R
\name{duplicated_count_index}
\alias{duplicated_count_index}
\title{Calculate count or Index of duplicated Values}
\usage{
duplicated_count_index(x, type = "count")
}
\arguments{
\item{x}{A data frame or a vector that can be coerced into a data frame.}

\item{type}{A character string specifying the type of calculation to perform. Possible values are "count" (default) or "index".}
}
\value{
If type = "count", it returns a vector with the count of duplicates for each row in the data frame x, excluding the row itself. If type = "index", it returns a vector with the index of duplicates for each row in the data frame x, where the index represents the number of previous duplicates (starting from 1) for each unique combination of values.
}
\description{
This function calculates either the count or index of duplicated values in a data frame.
}
\examples{
# Example 1: Count of duplicated rows
#data <- data.frame(a = c(1, 2, 3, 2, 1, 3, 4))
#duplicated_count_index(data, type = "count")

# Output:
# [1] 1 1 1 1 1 1 0

# Example 2: Index of duplicated rows
#data <- data.frame(a = c(1, 2, 3, 2, 1, 3, 4))
#duplicated_count_index(data, type = "index")

# Output:
# [1] 0 0 0 1 2 1 0
}
