% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_series.R
\name{fourier_series}
\alias{fourier_series}
\title{Fourier Series}
\usage{
fourier_series(x, n, period)
}
\arguments{
\item{x}{The input value for which the Fourier series is calculated.}

\item{n}{The number of terms in the Fourier series.}

\item{period}{The period of the Fourier series.}
}
\value{
The Fourier series expression as a character string.
}
\description{
This function calculates the Fourier series for a given input \code{x} with a specified number of terms \code{n} and period \code{period}.
}
\examples{
fourier_series(3, 5, 2)
# [1] "sin(3 * 1 * 2 * pi / 2) + cos(3 * 1 * 2 * pi / 2) + sin(3 * 2 * 2 * pi / 2) + cos(3 * 2 * 2 * pi / 2) + sin(3 * 3 * 2 * pi / 2) + cos(3 * 3 * 2 * pi / 2) + sin(3 * 4 * 2 * pi / 2) + cos(3 * 4 * 2 * pi / 2) + sin(3 * 5 * 2 * pi / 2) + cos(3 * 5 * 2 * pi / 2)"
}
