% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_attributes.R
\name{get_column_attributes}
\alias{get_column_attributes}
\title{Get column attributes}
\usage{
get_column_attributes(x, drop = c("class", "levels"))
}
\arguments{
\item{x}{The object for which attributes need to be retrieved.}

\item{drop}{A character vector specifying the names of attributes to be dropped from the result. Default is c("class", "levels").}
}
\value{
A list of attributes of the object, excluding the specified attributes to be dropped.
}
\description{
This function retrieves the attributes of a given object, excluding specified attributes to be dropped.
}
\examples{
# Example 1: Get attributes of a vector
vec <- c(1, 2, 3)
get_column_attributes(vec)

# Output:
# $dim
# NULL

}
