% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_significant_figures.R
\name{get_default_significant_figures}
\alias{get_default_significant_figures}
\title{Get the default number of significant figures}
\usage{
get_default_significant_figures(data)
}
\arguments{
\item{data}{\code{numeric(1)} A numerical vector}
}
\value{
If the data is numeric, "3", otherwise NA.
}
\description{
This function gets the default number of significant figures when given a numeric vector.
}
\examples{
x <- 1:10
get_default_significant_figures(x)
}
