% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_installed_packages_with_data.R
\name{get_installed_packages_with_data}
\alias{get_installed_packages_with_data}
\title{Get Installed Packages with Data}
\usage{
get_installed_packages_with_data(with_data = TRUE)
}
\arguments{
\item{with_data}{Logical value indicating whether to include only packages that contain data. Default is \code{TRUE}.}
}
\value{
A character vector containing the names of the installed packages. If \code{with_data} is \code{TRUE}, only packages with data will be included.
}
\description{
This function retrieves a list of installed packages in R, optionally including only those packages that contain data.
}
\examples{
# Get all installed packages
get_installed_packages_with_data()

# Get installed packages with data
get_installed_packages_with_data(with_data = TRUE)

# Get all installed packages (including those without data)
get_installed_packages_with_data(with_data = FALSE)

}
