% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lat_from_data.R
\name{get_lat_from_data}
\alias{get_lat_from_data}
\title{Get Latitude values from a Data file}
\usage{
get_lat_from_data(datafile)
}
\arguments{
\item{datafile}{A data file containing latitude values. The latitude values should be present in the first column, starting from the 5th row.}
}
\value{
A vector of unique latitude values extracted from the data file.
}
\description{
This function takes a data file as input and extracts the latitude values from it. The latitude values should be present in the first column of the data file, starting from the 5th row. The function removes any missing or non-numeric values and returns a vector of unique latitude values.
}
\examples{
# Example data file: mydata.csv
# Get latitude values from the data file
# data_file <- read.csv("mydata.csv")
# get_lat_from_data(data_file)

}
