% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lon_from_data.R
\name{get_lon_from_data}
\alias{get_lon_from_data}
\title{Get Longitude coordinates from Data}
\usage{
get_lon_from_data(datafile)
}
\arguments{
\item{datafile}{A data file containing longitude values. The longitude values should be present in the second column of the fifth row.}
}
\value{
A vector of unique longitude values extracted from the data file.
}
\description{
This function takes a data file as input and extracts the longitude values from it. The longitude values should be present in the second column of the fifth row. The function removes any missing or non-numeric values and returns a vector of unique longitude values.
The function returns a vector of unique longitude values extracted from the data file. It removes any missing or non-numeric values using the \code{na.omit()} function and transposes the extracted values using the \code{t()} function. The function is also exported, making it available for use outside the current package or script.
}
\examples{
# Example data file: mydata.csv
# Get longitude values from the data file
# data_file <- read.csv("mydata.csv")
# get_lon_from_data(data_file)

}
