% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quarter_label.R
\name{get_quarter_label}
\alias{get_quarter_label}
\title{Get Quarter of the year Label}
\usage{
get_quarter_label(quarter, start_month)
}
\arguments{
\item{quarter}{The quarter number. Must be in the range of 1 to 4.}

\item{start_month}{The starting month of the year. Must be in the range of 1 to 12.}
}
\value{
A factor object containing three-letter strings representing the specified quarters.
}
\description{
This function returns a three-letter string representing a specific quarter in a year.
}
\examples{
# Get quarter label starting from January
# get_quarter_label(quarter = 1, start_month = 1)

# Get quarter label starting from April
# get_quarter_label(quarter = 2, start_month = 4)

# Get quarter label starting from October
# get_quarter_label(quarter = 4, start_month = 10)

}
