% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_years_from_data.R
\name{get_years_from_data}
\alias{get_years_from_data}
\title{Get year from data}
\usage{
get_years_from_data(datafile)
}
\arguments{
\item{datafile}{A data file or data frame.}
}
\value{
A vector of unique years extracted from the data file.
}
\description{
Extracts unique years from a data file.
}
