% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashed_id.R
\name{hashed_id}
\alias{hashed_id}
\title{Hashed id}
\usage{
hashed_id(x, salt, algo = "crc32")
}
\arguments{
\item{x}{A vector or list of values to be hashed.}

\item{salt}{A character string representing a salt value to be appended to each element in x. Default is NULL, indicating no salt.}

\item{algo}{A character string specifying the hashing algorithm to be used. Default is "crc32".}
}
\value{
A character vector of hashed identifiers corresponding to the input values.
}
\description{
This function generates hashed identifiers by applying a specified algorithm to the input values, optionally incorporating a salt value.
}
\examples{
# Example 1: Generate hashed identifiers without salt
x <- c("apple", "banana", "cherry")
hashed_id(x)

# Output:
# [1] "AD5A2589" "77D3B209" "2C6EDC7A"

# Example 2: Generate hashed identifiers with salt
x <- c("apple", "banana", "cherry")
salt <- "salty"
hashed_id(x, salt)

# Output:
# [1] "1A1E6CE0E9A969A9" "1B69B1B8C1D6EF99" "1C10A25AA55025CD"
}
