% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_ODK.R
\name{import_from_ODK}
\alias{import_from_ODK}
\title{Import from ODK}
\usage{
import_from_ODK(username, form_name, platform)
}
\arguments{
\item{username}{Your username (character) on the ODK platform.}

\item{form_name}{The name of the form (character) you want to import}

\item{platform}{The ODK platform (character)you are using. Valid options are "kobo" or "ona".}
}
\value{
The imported form data as a structured object.
}
\description{
This function imports data from ODK platforms, such as KoboToolbox or Ona, based on the specified username, form name, and platform.
}
