% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.binary.R
\name{is.binary}
\alias{is.binary}
\title{Checking Binary variable}
\usage{
is.binary(x)
}
\arguments{
\item{x}{The input variable to be checked for binary nature.}
}
\value{
The function returns a logical value indicating whether the input is binary or not. It returns TRUE if the input is binary, and FALSE otherwise.
}
\description{
This function checks if the input is a binary variable. It determines whether the input is of logical, numeric, or factor type and checks if it meets the criteria for a binary variable.
}
\examples{
is.binary(TRUE)           # TRUE
is.binary(c(0, 1, 1, 0))  # TRUE
is.binary(factor(c("Yes", "No", "Yes", "Yes")))  # TRUE
is.binary(c(1, 2, 3, 4))  # FALSE
 
}
