% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.containValueLabel.R
\name{is.containValueLabel}
\alias{is.containValueLabel}
\title{Checking if an object contains a value label attribute}
\usage{
is.containValueLabel(x)
}
\arguments{
\item{x}{The object to be checked for the presence of a value label attribute.}
}
\value{
The function returns a logical value indicating whether the input object contains a value label attribute. It returns TRUE if the attribute is present, and FALSE otherwise.
}
\description{
This function checks if an object contains a value label attribute. It determines whether the input object has an attribute with the name "labels_label".
}
\examples{
df <- data.frame(x = 1:10, y = 11:20)
attributes(df)$labels_label <- "Value Labels"
is.containValueLabel(df)  # TRUE

vec <- c(1, 2, 3, 4, 5)
is.containValueLabel(vec)  # FALSE 

}
