% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.emptyvariable.R
\name{is.emptyvariable}
\alias{is.emptyvariable}
\title{Check if Variable is Empty}
\usage{
is.emptyvariable(x)
}
\arguments{
\item{x}{The variable to be checked.}
}
\value{
A logical value indicating whether the variable is empty or not.
}
\description{
This function checks if a variable is empty, i.e., if it contains only empty values.
}
\examples{
is.emptyvariable(c("", "abc", ""))
# [1] FALSE

is.emptyvariable(c("", ""))
# [1] TRUE
}
