% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelscount.R
\name{is.levelscount}
\alias{is.levelscount}
\title{Check Number of Levels in a Factor}
\usage{
is.levelscount(x, n)
}
\arguments{
\item{x}{The factor variable to be checked.}

\item{n}{The expected number of levels.}
}
\value{
A logical value indicating whether the factor variable has the specified number of levels.
}
\description{
This function checks if a factor variable has a specific number of levels.
}
\examples{
#' # Example 1: Validating a factor variable with the correct number of levels
#factor_var <- factor(c("A", "B", "C", "D"))
#is.levelscount(factor_var, 4)
# Expected output: TRUE

# Example 2: Validating a factor variable with an incorrect number of levels
#factor_var <- factor(c("A", "B", "C", "D"))
#is.levelscount(factor_var, 3)
# Expected output: FALSE

}
