% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.logical.like.R
\name{is.logical.like}
\alias{is.logical.like}
\title{Logical like}
\usage{
is.logical.like(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
TRUE if the object is logical-like, FALSE otherwise.
}
\description{
Check if an object is logical-like.
}
\examples{
is.logical.like(TRUE) 
# Output: TRUE

is.logical.like(c(TRUE, FALSE)) 
# Output: TRUE

is.logical.like(1) 
# Output: TRUE

is.logical.like(c(0, 1)) 
# Output: TRUE

is.logical.like("TRUE") 
# Output: FALSE

is.logical.like(NULL) 
# Output: FALSE
}
