% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.NAvariable.R
\name{is.NAvariable}
\alias{is.NAvariable}
\title{Is NA variable}
\usage{
is.NAvariable(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
TRUE if the variable consists entirely of NA values, FALSE otherwise.
}
\description{
Check if a variable consists entirely of NA values.
}
\examples{
is.NAvariable(c(NA, NA, NA))
# Output: TRUE

is.NAvariable(c(1, 2, 3))
# Output: FALSE

is.NAvariable(c(TRUE, FALSE, NA))
# Output: FALSE

is.NAvariable(NULL)
# Output: FALSE
}
