% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_lon_dataframe.R
\name{lat_lon_dataframe}
\alias{lat_lon_dataframe}
\title{Create a dataframe with latitude and longitude information from a data file.}
\usage{
lat_lon_dataframe(datafile)
}
\arguments{
\item{datafile}{The path or name of the data file.}
}
\value{
A data frame containing latitude, longitude, and station information.
}
\description{
This function creates a data frame with latitude and longitude information from a data file.
}
\examples{
# Create a latitude-longitude data frame
# datafile <- "data.csv"
# lat_lon_dataframe(datafile) 

}
