% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_get_dim_min_max.R
\name{nc_get_dim_min_max}
\alias{nc_get_dim_min_max}
\title{Retrieve minimum and maximum values of a dimension from a NetCDF file.}
\usage{
nc_get_dim_min_max(nc, dimension, time_as_date = TRUE)
}
\arguments{
\item{nc}{A NetCDF file object.}

\item{dimension}{The name of the dimension for which to retrieve the minimum and maximum values.}

\item{time_as_date}{A logical value indicating whether to treat time dimension values as dates. Default is TRUE.}
}
\value{
A numeric vector containing the minimum and maximum values of the dimension.
}
\description{
Retrieves the minimum and maximum values of a dimension from a NetCDF file.
}
\examples{
# nc_file <- nc_open("path/to/netcdf/file.nc")
# min_max <- nc_get_dim_min_max(nc_file, "time", time_as_date = TRUE)
# nc_close(nc_file)


}
