% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_default_item.R
\name{next_default_item}
\alias{next_default_item}
\title{Title}
\usage{
next_default_item(
  prefix,
  existing_names = c(),
  include_index = FALSE,
  start_index = 1
)
}
\arguments{
\item{prefix}{}

\item{existing_names}{}

\item{include_index}{}

\item{start_index}{}
}
\description{
Title
}
