% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_CPT.R
\name{output_CPT}
\alias{output_CPT}
\title{Generate CPT data for visualization}
\usage{
output_CPT(
  data,
  lat_lon_data,
  station_latlondata,
  latitude,
  longitude,
  station,
  year,
  element,
  long.data = TRUE,
  na_code = -999
)
}
\arguments{
\item{data}{The input data frame containing the raw data.}

\item{lat_lon_data}{The data frame containing latitude and longitude information.}

\item{station_latlondata}{The column name representing the station in the lat_lon_data data frame.}

\item{latitude}{The column name representing latitude in the data frame.}

\item{longitude}{The column name representing longitude in the data frame.}

\item{station}{The column name representing the station in the data frame.}

\item{year}{The column name representing the year in the data frame.}

\item{element}{The column name representing the element in the data frame.}

\item{long.data}{If TRUE, the data frame contains all the required variables. If FALSE, the data frame contains multiple years stacked as columns.}

\item{na_code}{The code to be used for missing values. Default is -999.}
}
\value{
The CPT data frame for visualization.
}
\description{
This function generates a CPT (Color Palette Table) data frame for visualization purposes. It prepares the data by rearranging and merging the necessary variables and information. The resulting CPT data frame can be used for visualizing spatial data using CPT-based software or libraries.
}
\examples{
# Example usage
# data <- read.csv("data.csv")
# lat_lon_data <- read.csv("lat_lon_data.csv")
# output_CPT(data, lat_lon_data, "station_latlon", "latitude", "longitude", "station", "year", "element")

# Example usage with additional parameters
# output_CPT(data, lat_lon_data, "station_latlon", "latitude", "longitude", "station", "year", "element", long.data = FALSE, na_code = -1)

}
