% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_check.R
\name{package_check}
\alias{package_check}
\title{Package Check}
\usage{
package_check(package)
}
\arguments{
\item{package}{A character string specifying the name of the package to be checked.}
}
\value{
A numeric vector with the following elements:
\itemize{
\item \code{[1]} Status code:
\itemize{
\item \code{0} - Package not found (incorrect spelling).
\item \code{1} - Package found and installed.
\item \code{2} - Package found, but not installed.
}
\item \code{[2]} Version comparison result:
\itemize{
\item \code{-1} - Installed version is older than the latest version.
\item \code{0} - Installed version is the same as the latest version.
\item \code{1} - Installed version is newer than the latest version.
}
\item \code{[3]} Installed version (if available), as a character string.
\item \code{[4]} Latest version available online (if available), as a character string.
}
}
\description{
This function checks the status of a specified package in the current R environment. It verifies whether the package is installed, and if so, it compares the installed version with the latest version available online.
}
\examples{

# Check package "dplyr"
#package_check("dplyr")

# Check package "ggplot2"
#package_check("ggplot2")
}
