% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pentad.R
\name{pentad}
\alias{pentad}
\title{Calculate the pentad for a given date}
\usage{
pentad(date)
}
\arguments{
\item{date}{The input date.}
}
\value{
The pentad number corresponding to the input date.
}
\description{
This function calculates the pentad (5-day period) for a given date. It determines which pentad the date falls into based on the day of the month.
}
\examples{
# Example usage
# pentad(as.Date("2023-07-17"))
 
}
