% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_declustered.R
\name{plot_declustered}
\alias{plot_declustered}
\title{Generate declustering plots}
\usage{
plot_declustered(
  data,
  station_col_name,
  element_col_name,
  threshold,
  r = NULL,
  xlab = NULL,
  ylab = NULL,
  ncol = 1,
  print_summary = FALSE
)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{station_col_name}{The name of the column in the data frame that represents the stations.}

\item{element_col_name}{The name of the column in the data frame that represents the elements.}

\item{threshold}{The threshold value used for declustering.}

\item{r}{The run length parameter used for declustering. Default is NULL.}

\item{xlab}{The label for the x-axis of the plot. Default is NULL.}

\item{ylab}{The label for the y-axis of the plot. Default is NULL.}

\item{ncol}{The number of columns in the plot grid. Default is 1.}

\item{print_summary}{If TRUE, it prints the declustering summary for each station. Default is FALSE.}
}
\value{
If print_summary is FALSE, it returns a plot object(s). Otherwise, it returns NULL.
}
\description{
This function generates declustering plots based on the provided data and parameters. It uses the texmex package for declustering analysis.
}
\examples{
# Example usage
# data <- read.csv("data.csv")
# plot_declustered(data, "station", "element", threshold = 0.5)

# Example usage with additional parameters
# plot_declustered(data, "station", "element", threshold = 0.5, r = 10, xlab = "Time", ylab = "Value", ncol = 2, print_summary = TRUE)

}
