% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mrl.R
\name{plot_mrl}
\alias{plot_mrl}
\title{Plot Mean Residual Life}
\usage{
plot_mrl(
  data,
  station_name,
  element_name,
  umin,
  umax,
  ncol = 1,
  xlab = "Threshold",
  ylab = "Mean excess",
  fill = "red",
  col = "black",
  rug = TRUE,
  addNexcesses = TRUE,
  textsize = 4
)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{station_name}{The name of the column in \code{data} representing the stations.}

\item{element_name}{The name of the column in \code{data} representing the elements.}

\item{umin}{The lower threshold value. If not provided, the minimum value in \code{element_name} will be used.}

\item{umax}{The upper threshold value. If not provided, the maximum value in \code{element_name} will be used.}

\item{ncol}{The number of columns for arranging the subplots. Default is 1.}

\item{xlab}{The label for the x-axis. Default is "Threshold".}

\item{ylab}{The label for the y-axis. Default is "Mean excess".}

\item{fill}{The fill color for the plot. Default is "red".}

\item{col}{The color for the plot. Default is "black".}

\item{rug}{A logical value indicating whether to include rug plot. Default is TRUE.}

\item{addNexcesses}{A logical value indicating whether to add the number of excesses. Default is TRUE.}

\item{textsize}{The size of the text in the plot. Default is 4.}
}
\value{
If \code{station_name} is provided, a grid of plots is returned. Otherwise, a single plot is returned.
}
\description{
This function generates a plot of the mean excess for a given threshold in a dataset.
}
\examples{
# Example 1: Single plot
# Generate a data frame
data <- data.frame(
  station = c("A", "A", "B", "B", "C", "C"),
  element = c(10, 15, 20, 25, 30, 35)
)

# Generate a plot for the element column with default settings
plot_mrl(data, element_name = "element")

# Example 2: Grid of plots
# Generate a data frame
data <- data.frame(
  station = c("A", "A", "B", "B", "C", "C"),
  element = c(10, 15, 20, 25, 30, 35)
)

# Generate a grid of plots for each station
plot_mrl(data, station_name = "station", element_name = "element", ncol = 2)

}
