% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{Slope}
\usage{
slope(y, x)
}
\arguments{
\item{y}{A numeric vector of response variable values.}

\item{x}{A numeric vector of predictor variable values.}
}
\value{
The slope of the linear regression model.
}
\description{
Calculate the slope of a linear regression model.
}
\examples{
y <- c(1, 2, 3, 4, 5)
x <- c(2, 4, 6, 8, 10)
slope(y, x)

}
