% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spei_input.R
\name{spei_input}
\alias{spei_input}
\title{Calculate SPEI Input Data}
\usage{
spei_input(data, station, year, month, element)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{The name of the year column in \code{data}.}

\item{month}{The name of the month column in \code{data}.}

\item{element}{The name of the column(s) in \code{data} to apply the condition to.}
}
\value{
A time series object (\code{ts}) containing the calculated SPEI input data.
}
\description{
This function calculates the Standardized Precipitation-Evapotranspiration Index (SPEI) input data from a given data frame. The function sorts the data by year and month, checks for data completeness, and prepares the data for further SPEI calculations.

#' @details
The function expects the input data to be in a data frame format. It calculates the SPEI input data by performing the following steps:
\enumerate{
\item Sorts the data frame by the year and month columns in ascending order.
\item Verifies if the sorted data frame matches the original data frame, ensuring correct sorting for SPEI/SPI calculations. If they differ, an error is raised.
\item Checks if there are multiple values per month (per station) in the data frame. If multiple values are detected, an error is raised as SPEI/SPI requires monthly data with one value per month.
\item If the \code{station} parameter is provided, the function performs additional checks for each unique station:
\itemize{
\item Filters the data frame to include only rows with the current station.
\item Generates a sequence of dates from the first date in the filtered data to the last date, incrementing by one month.
\item Compares the length of the generated date sequence with the number of rows in the filtered data. If they differ, an error is raised, indicating missing months in the data for the current station.
}
\item Constructs the \code{cols} variable by combining the \code{id_cols} and \code{element} columns.
\item Determines the start year and month based on the first values in the \code{year} and \code{month} columns.
\item If the \code{station} parameter is provided, the function transforms the data frame into a "wide" format using \code{pivot_wider}, organizing the data by year and month with station-specific columns. Missing values are filled with \code{NA}.
\item Converts the resulting data frame to a time series object (\code{ts}) using \code{as.matrix}. The frequency is set to 12 for monthly data, and the start year and month are determined from the data.
\item If the \code{station} parameter is not provided, the function directly converts the \code{element} column to a time series object (\code{ts}) using \code{as.matrix}, with a frequency of 12 and the determined start year and month.
\item Returns the calculated time series data as the output.
}

#' @keywords
SPEI, precipitation, evapotranspiration, time series, data preparation
}
\examples{
# TODO


}
