% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spei_output.R
\name{spei_output}
\alias{spei_output}
\title{Extract SPEI/SPI Column from 'spei' Object}
\usage{
spei_output(x, data, station, year, month)
}
\arguments{
\item{x}{An object of class 'spei'.}

\item{data}{The data.frame to calculate from.}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{The name of the year column in \code{data}.}

\item{month}{The name of the month column in \code{data}.}
}
\value{
A vector containing the extracted SPEI/SPI column from the 'spei' object 'x'.
}
\description{
This function extracts the Standardized Precipitation-Evapotranspiration Index (SPEI) or Standardized Precipitation Index (SPI) column from a 'spei' object. It is designed to work with the original data and handle multiple stations if present. The function removes NA values introduced when unstacking the data to return a vector of the correct length.
}
\examples{
# TODO:



}
