% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_items_in_groups.R
\name{split_items_in_groups}
\alias{split_items_in_groups}
\title{Split items into groups}
\usage{
split_items_in_groups(items, num)
}
\arguments{
\item{items}{A vector of items to be split into groups.}

\item{num}{The number of groups to split the items into.}
}
\value{
A list containing the groups of items. Each element of the list represents a group and contains a subset of the original items.
}
\description{
This function takes a vector of items and splits them into a specified number of groups. The number of items must be divisible by the number of groups, otherwise an error is thrown.
}
\examples{
 items <- c("A", "B", "C", "D", "E", "F", "G", "H")
num_groups <- 2
split_items_in_groups(items, num_groups)

# Output:
# [[1]]
# [1] "A" "B" "C" "D"
#
# [[2]]
# [1] "E" "F" "G" "H"
}
