% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sample.R
\name{summary_sample}
\alias{summary_sample}
\title{#' summary sample}
\usage{
summary_sample(x, size, replace = FALSE)
}
\arguments{
\item{x}{A vector from which to draw a sample.}

\item{size}{The size of the sample to be drawn.}

\item{replace}{Logical indicating whether sampling should be done with replacement (default: FALSE).}
}
\value{
A summary of the sample drawn from the input vector.
}
\description{
A function that summarizes a sample from a vector.
}
\examples{
# Example 1: Draw a sample from a vector
data <- c(1, 2, 3, 4, 5)
sample <- summary_sample(x = data, size = )

# Example 2: Draw a sample from a vector with replacement
data <- c("A", "B", "C", "D", "E")
sample <- summary_sample(x = data, size = 4, replace = TRUE)
}
