% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_Plot.R
\name{threshold_Plot}
\alias{threshold_Plot}
\title{Threshold Plot}
\usage{
threshold_Plot(
  x,
  r,
  type = c("GP", "PP", "Exponential"),
  nint = 10,
  alpha = 0.05,
  na.action = na.omit,
  xlb = "",
  main = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Numeric vector of data values.}

\item{r}{Numeric vector of quantiles defining the range of thresholds. Default is to use the 75th and 99th percentiles of x.}

\item{type}{Character string specifying the type of extreme value distribution to fit. Possible values are "GP" , "PP" , or "Exponential". Default is "GP".}

\item{nint}{Integer specifying the number of intervals between the lower and upper thresholds. Default is 10.}

\item{alpha}{Numeric value specifying the significance level for confidence intervals. Default is 0.05.}

\item{na.action}{Function to handle missing values. Default is na.omit.}

\item{xlb}{Character string specifying the x-axis label for the plot. Default is an empty string.}

\item{main}{Character string specifying the main title for the plot. Default is NULL.}

\item{verbose}{Logical value indicating whether to print additional output. Default is FALSE.}

\item{...}{Additional arguments to be passed to the fitting and confidence interval functions.}
}
\value{
A list of threshold plots, including location, scale, and shape, depending on the specified type.
}
\description{
This function generates threshold plots for a given dataset.
}
\examples{
# Generate threshold plots for a dataset
data <- c(2.1, 2.2, 3.4, 4.5, 5.6, 6.7, 7.8, 8.9, 9.0, 10.1)
threshold_Plot(data, type = "GP")
}
