% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_pollution_rose.R
\name{wind_pollution_rose}
\alias{wind_pollution_rose}
\title{Wind pollution Rose}
\usage{
wind_pollution_rose(
  mydata,
  date_name,
  pollutant,
  type1_col_name,
  type2_col_name,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing the data for the plot.}

\item{date_name}{The name of the column that contains the date information.}

\item{pollutant}{The name of the column that contains the pollutant information.}

\item{type1_col_name}{The name of the first type column.}

\item{type2_col_name}{The name of the second type column.}

\item{...}{Additional arguments to be passed to the underlying plotting functions.}
}
\value{
The generated wind or pollution rose plot.
}
\description{
This function creates a wrapper around windRose and pollutionRose functions from openair package
}
\examples{
# Example 1: Creating a wind rose plot
# wind_pollution_rose(mydata = wind_data, date_name = "date")
}
