% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwr_report.R
\name{wwr_export}
\alias{wwr_export}
\title{Reshape data into formats required by WMO for submission of climatic data}
\usage{
wwr_export(
  data,
  year,
  month,
  mean_station_pressure,
  mean_sea_level_pressure,
  mean_temp,
  total_precip,
  mean_max_temp,
  mean_min_temp,
  mean_rel_hum,
  link,
  link_by,
  station_data,
  wmo_number,
  latitude,
  longitude,
  country_name,
  station_name,
  height_station,
  height_barometer,
  wigos_identifier,
  folder
)
}
\arguments{
\item{data}{The input data frame containing the climatic data.}

\item{year}{The name of the column in 'data' representing the year.}

\item{month}{The name of the column in 'data' representing the month.}

\item{mean_station_pressure}{The name of the column in 'data' representing the mean station pressure.}

\item{mean_sea_level_pressure}{The name of the column in 'data' representing the mean sea level pressure.}

\item{mean_temp}{The name of the column in 'data' representing the mean daily air temperature.}

\item{total_precip}{The name of the column in 'data' representing the total precipitation.}

\item{mean_max_temp}{The name of the column in 'data' representing the mean daily maximum air temperature.}

\item{mean_min_temp}{The name of the column in 'data' representing the mean daily minimum air temperature.}

\item{mean_rel_hum}{The name of the column in 'data' representing the mean of the daily relative humidity.}

\item{link}{The name of the column in 'data' representing the link between the data and station information.}

\item{link_by}{The method for linking the data and station information. Possible values: "wmo_number", "station_name".}

\item{station_data}{The data frame containing the station information.}

\item{wmo_number}{The name of the column in 'station_data' representing the WMO number.}

\item{latitude}{The name of the column in 'station_data' representing the latitude.}

\item{longitude}{The name of the column in 'station_data' representing the longitude.}

\item{country_name}{The name of the column in 'station_data' representing the country name.}

\item{station_name}{The name of the column in 'station_data' representing the station name.}

\item{height_station}{The name of the column in 'station_data' representing the station height.}

\item{height_barometer}{The name of the column in 'station_data' representing the barometer height.}

\item{wigos_identifier}{The name of the column in 'station_data' representing the WIGOS station identifier.}

\item{folder}{The folder where the output files will be saved.}
}
\value{
TODO
}
\description{
The function is meant to reshape data into formats required by WMO for submission of climatic data.This gives Yearly data records with monthly and annual data for a particular year:
}
\examples{
# TODO

}
