% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_unused_levels.R
\name{drop_unused_levels}
\alias{drop_unused_levels}
\title{Drop Unused Levels}
\usage{
drop_unused_levels(dat, columns)
}
\arguments{
\item{columns}{A character vector specifying the column names in \code{dat} from which unused levels should be dropped.}

\item{data}{A data frame.}
}
\value{
A modified version of the input data frame \code{data} with unused levels dropped from the specified columns.
}
\description{
This function drops unused levels from specified columns in a data frame.
}
\examples{
# Create a data frame
df <- data.frame(A = factor(c("apple", "banana", "apple", "orange")),
                 B = factor(c("red", "blue", "green", "red")),
                 C = c(1, 2, 3, 4))

# Drop unused levels from column 'A'
drop_unused_levels(df, columns = "A")
# Output:
#       A    B C
# 1 apple  red 1
# 2 banana blue 2
# 3 apple green 3
# 4 orange  red 4

# Drop unused levels from multiple columns
drop_unused_levels(df, columns = c("A", "B"))
# Output:
#       A    B C
# 1 apple  red 1
# 2 banana blue 2
# 3 apple green 3
# 4 orange  red 4
}
