% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_years_from_data.R
\name{get_years_from_data}
\alias{get_years_from_data}
\title{Get year from data}
\usage{
get_years_from_data(datafile)
}
\arguments{
\item{datafile}{A data file or data frame.}
}
\value{
A vector of unique years extracted from the data file.
}
\description{
Extracts unique years from a data file.
}
\examples{
# # Sample data file
# datafile <- data.frame(
#   Name = c("John", "Alice", "Bob"),
#   Age = c(25, 30, 35),
#   "2019" = c(100, 150, 200),
#   "2020" = c(200, 300, 400),
#   "2021" = c(300, 450, 600)
# )
#
# # Get years from data file
# years <- get_years_from_data(datafile)
# years
#
}
